transformed updateReality to hexagonal architecture
