+++ src/main/java/com/example/demo/domain/ports/UpdateRealitiesInputPort.java
@@ -1,0 +1,7 @@
+package com.example.demo.domain.ports;
+
+import com.example.demo.domain.models.Reality;
+
+public interface UpdateRealitiesInputPort {
+    void updateReality(Reality reality, Long realityId);
+}
